feat: add cjsToEsm  metod
